requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
decorator==4.3.0
Flask==1.0.2
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
MarkupSafe==1.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.2.0
python-chess==0.23.10
requests==2.20.1
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
